+++ .Dev-App/Unity/Assets/Demo/AddressablesImpl/AddressablesProviderFactory.cs
@@ -1,9 +1,17 @@
-using GuiToolkit;
 using GuiToolkit.AssetHandling;
 using UnityEngine;
 
-[CreateAssetMenu(fileName = nameof(AddressablesProviderFactory), menuName = "AddressablesProviderFactory")]
+/// <summary>
+/// Factory for creating <see cref="AddressablesAssetProvider"/> instances.
+/// Can be added to the project via the Unity "Create" menu.
+/// </summary>
+[CreateAssetMenu(
+    fileName = nameof(AddressablesProviderFactory),
+    menuName = "AddressablesProviderFactory")]
 public class AddressablesProviderFactory : AbstractAssetProviderFactory
 {
-	public override IAssetProvider CreateProvider() => new AddressablesProvider();
+    /// <summary>
+    /// Creates a new <see cref="AddressablesAssetProvider"/> instance.
+    /// </summary>
+    public override IAssetProvider CreateProvider() => new AddressablesAssetProvider();
 }
