+++ Runtime/Code/Basic/AssetHandling/AssetLoader.cs
@@ -1,122 +1,147 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using GuiToolkit.Exceptions;
 using UnityEngine;
 
 namespace GuiToolkit.AssetHandling
 {
-	public class AssetLoader
-	{
-		public static AssetLoader s_instance;
-		private class LoadedPrefab
-		{
-			public GameObject Prefab;
-			public IInstanceHandle Handle;
-		}
+    /// <summary>
+    /// High-level loader that creates UiPanels either directly or via pool.
+    /// Delegates key resolution and instantiation to AssetManager / UiPool.
+    /// </summary>
+    public class AssetLoader
+    {
+        public static AssetLoader s_instance;
 
-		private readonly HashSet<string> m_BeingLoaded = new();
-		private readonly Dictionary<string, LoadedPrefab> m_loadedPrefabsByClassName = new();
+        private class LoadedPrefab
+        {
+            public GameObject Prefab;
+            public IInstanceHandle Handle;
+        }
 
+        private readonly System.Collections.Generic.HashSet<string> m_BeingLoaded = new();
+        private readonly System.Collections.Generic.Dictionary<string, LoadedPrefab> m_loadedPrefabsByClassName = new();
 
-		public static AssetLoader Instance
-		{
-			get
-			{
-				if (s_instance == null)
-					s_instance = new AssetLoader();
-				return s_instance;
-			}
-		}
+        /// <summary>
+        /// Singleton instance of the loader.
+        /// </summary>
+        public static AssetLoader Instance
+        {
+            get
+            {
+                if (s_instance == null)
+                    s_instance = new AssetLoader();
 
-		public void LoadAsync( UiPanelLoadInfo _loadInfo )
-		{
-			if (_loadInfo == null)
-				throw new ArgumentNullException(nameof(_loadInfo));
+                return s_instance;
+            }
+        }
 
-			if (_loadInfo.MaxInstances > 0)
-			{
-				int openInstances = UiPanel.GetNumOpenDialogs(_loadInfo.PanelType);
-				UiPanel.DebugLogStatic($"Checking max instances {_loadInfo.PanelType.Name}: openInstances:{openInstances} loadInfo.MaxInstances:{_loadInfo.MaxInstances}");
-				if (openInstances >= _loadInfo.MaxInstances)
-				{
-					Debug.LogWarning($"Attempt to load panel '{_loadInfo.PanelType.Name}', " +
-									 $"but open instances ({openInstances}) is >= allowed instances in load info {_loadInfo.MaxInstances}, cancelling load without callbacks");
-					return;
-				}
-			}
+        /// <summary>
+        /// Load a panel according to the provided load info.
+        /// Decides between direct instantiate and pool path.
+        /// </summary>
+        /// <param name="_loadInfo">Panel load parameters.</param>
+        public void LoadAsync(UiPanelLoadInfo _loadInfo)
+        {
+            if (_loadInfo == null)
+                throw new ArgumentNullException(nameof(_loadInfo));
 
-			if (_loadInfo.InstantiationType == UiPanelLoadInfo.EInstantiationType.Pool)
-			{
-				LoadAsyncFromPool(_loadInfo);
-				return;
-			}
+            if (_loadInfo.MaxInstances > 0)
+            {
+                int openInstances = UiPanel.GetNumOpenDialogs(_loadInfo.PanelType);
+                UiPanel.DebugLogStatic(
+                    $"Checking max instances {_loadInfo.PanelType.Name}: openInstances:{openInstances} loadInfo.MaxInstances:{_loadInfo.MaxInstances}"
+                );
 
-			LoadAsyncImpl(_loadInfo);
-		}
+                if (openInstances >= _loadInfo.MaxInstances)
+                {
+                    Debug.LogWarning(
+                        $"Attempt to load panel '{_loadInfo.PanelType.Name}', " +
+                        $"but open instances ({openInstances}) is >= allowed instances in load info {_loadInfo.MaxInstances}, cancelling load without callbacks"
+                    );
+                    return;
+                }
+            }
 
-		private async Task LoadAsyncImpl( UiPanelLoadInfo _loadInfo )
-		{
-			try
-			{
-				var key = new CanonicalAssetKey(_loadInfo.AssetProvider, _loadInfo.CanonicalId, _loadInfo.PanelType);
+            if (_loadInfo.InstantiationType == UiPanelLoadInfo.EInstantiationType.Pool)
+            {
+                LoadAsyncFromPool(_loadInfo);
+                return;
+            }
 
-				var instHandle = await AssetManager.InstantiateAsync(key, _loadInfo.Parent);
-				var go = instHandle.Instance;
-				var panel = go != null ? go.GetComponent(_loadInfo.PanelType) as UiPanel : null;
+            LoadAsyncImpl(_loadInfo);
+        }
 
-				if (panel == null)
-				{
-					instHandle.Release();
-					throw new AssetLoadFailedException(key,
-						$"Instance does not contain requested panel type '{_loadInfo.PanelType.Name}'.");
-				}
+        // =====================================================================
+        // private helpers
+        // =====================================================================
 
-				if (_loadInfo.InitPanelData != null)
-					panel.Init(_loadInfo.InitPanelData, instHandle);
+        private async Task LoadAsyncImpl(UiPanelLoadInfo _loadInfo)
+        {
+            try
+            {
+                var key = new CanonicalAssetKey(_loadInfo.AssetProvider, _loadInfo.CanonicalId, _loadInfo.PanelType);
 
-				_loadInfo.OnSuccess?.Invoke(panel);
-			}
-			catch (OperationCanceledException ex)
-			{
-				_loadInfo.OnFail?.Invoke(_loadInfo, ex);
-			}
-			catch (Exception ex)
-			{
-				_loadInfo.OnFail?.Invoke(_loadInfo, ex);
-			}
-		}
+                var instHandle = await AssetManager.InstantiateAsync(key, _loadInfo.Parent);
+                var go = instHandle.Instance;
+                var panel = go != null ? go.GetComponent(_loadInfo.PanelType) as UiPanel : null;
 
-		private async Task LoadAsyncFromPool( UiPanelLoadInfo _loadInfo )
-		{
-			try
-			{
-				var key = new CanonicalAssetKey(_loadInfo.AssetProvider, _loadInfo.CanonicalId, _loadInfo.PanelType);
+                if (panel == null)
+                {
+                    instHandle.Release();
+                    throw new AssetLoadFailedException(
+                        key,
+                        $"Instance does not contain requested panel type '{_loadInfo.PanelType.Name}'."
+                    );
+                }
 
-				var go = await UiPool.Instance.GetAsync(key);
-				var panel = go != null ? go.GetComponent(_loadInfo.PanelType) as UiPanel : null;
+                if (_loadInfo.InitPanelData != null)
+                    panel.Init(_loadInfo.InitPanelData, instHandle);
 
-				if (panel == null)
-				{
-					UiPool.Instance.Release(go);
-					throw new AssetLoadFailedException(key,
-						$"Pooled instance does not contain requested panel type '{_loadInfo.PanelType.Name}'.");
-				}
+                _loadInfo.OnSuccess?.Invoke(panel);
+            }
+            catch (OperationCanceledException ex)
+            {
+                _loadInfo.OnFail?.Invoke(_loadInfo, ex);
+            }
+            catch (Exception ex)
+            {
+                _loadInfo.OnFail?.Invoke(_loadInfo, ex);
+            }
+        }
 
-				if (_loadInfo.InitPanelData != null)
-					panel.Init(_loadInfo.InitPanelData, null);
+        private async Task LoadAsyncFromPool(UiPanelLoadInfo _loadInfo)
+        {
+            try
+            {
+                var key = new CanonicalAssetKey(_loadInfo.AssetProvider, _loadInfo.CanonicalId, _loadInfo.PanelType);
 
-				_loadInfo.OnSuccess?.Invoke(panel);
-			}
-			catch (OperationCanceledException ex)
-			{
-				_loadInfo.OnFail?.Invoke(_loadInfo, ex);
-			}
-			catch (Exception ex)
-			{
-				_loadInfo.OnFail?.Invoke(_loadInfo, ex);
-			}
-		}
-	}
+                var go = await UiPool.Instance.GetAsync(key);
+                var panel = go != null ? go.GetComponent(_loadInfo.PanelType) as UiPanel : null;
+
+                if (panel == null)
+                {
+                    UiPool.Instance.Release(go);
+                    throw new AssetLoadFailedException
+                    (
+                        key,
+                        $"Pooled instance does not contain requested panel type '{_loadInfo.PanelType.Name}'."
+                    );
+                }
+
+                if (_loadInfo.InitPanelData != null)
+                    panel.Init(_loadInfo.InitPanelData, null);
+
+                _loadInfo.OnSuccess?.Invoke(panel);
+            }
+            catch (OperationCanceledException ex)
+            {
+                _loadInfo.OnFail?.Invoke(_loadInfo, ex);
+            }
+            catch (Exception ex)
+            {
+                _loadInfo.OnFail?.Invoke(_loadInfo, ex);
+            }
+        }
+    }
 }
